+++ gwemopt/io/__init__.py
@@ -2,4 +2,4 @@ from gwemopt.io.export_tiles import (
     export_tiles_coverage_hist,
     export_tiles_coverage_int,
 )
-from gwemopt.io.schedule import read_schedule
+from gwemopt.io.schedule import export_schedule_xml, read_schedule, summary
